게시물 테이블 구조 변경, memberId 추가, doWrite할 때 로그인 체크, 작성자 정보 저장

+++ src/main/java/com/KoreaIT/kjs/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.KoreaIT.kjs.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -28,11 +30,21 @@ public class UsrArticleController {
 	// 액션 메서드
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
+	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
 		
+		Integer memberId = (Integer) httpSession.getAttribute("loginedMemberId");
+
+		if (memberId == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요.");
+		}
+
 		Article article = articleService.getArticle(id);
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다.", id), id);
+			return ResultData.from("F-2", Ut.f("%d번 글은 존재하지 않습니다.", id), id);
+		}
+		
+		if (memberId != article.getMemberId()) {
+			return ResultData.from("F-3", "수정 권한이 없습니다.");
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -42,11 +54,21 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+
+		Integer memberId = (Integer) httpSession.getAttribute("loginedMemberId");
+
+		if (memberId == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요.");
+		}
 		
 		Article article = articleService.getArticle(id);
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다.", id), id);
+			return ResultData.from("F-2", Ut.f("%d번 글은 존재하지 않습니다.", id), id);
+		}
+		
+		if (memberId != article.getMemberId()) {
+			return ResultData.from("F-3", "삭제 권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
@@ -56,15 +78,22 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession httpSession, String title, String body) {
+		
+		Integer loginedMemberId = (Integer) httpSession.getAttribute("loginedMemberId");
+		
+		if (loginedMemberId == null) {
+			return ResultData.from("F-1", "로그인 후 이용해주세요.");
+		}
+		
 		if (Ut.empty(title)) {
-			return ResultData.from("F-1", "제목을 입력해주세요.");
+			return ResultData.from("F-2", "제목을 입력해주세요.");
 		}
 		if (Ut.empty(title)) {
-			return ResultData.from("F-2", "내용을 입력해주세요.");
+			return ResultData.from("F-3", "내용을 입력해주세요.");
 		}
-
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
+		
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
 		
 		int id = (int) writeArticleRd.getData1();
 		Article article = articleService.getArticle(id);
